backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
python-multipart==0.0.7
httpx==0.26.0
pydantic-settings==2.1.0
google-generativeai>=0.8.0
python-dotenv==1.0.0
